+++ dialect/psql/table.go
@@ -9,6 +9,7 @@ import (
 	"github.com/stephenafamo/bob/dialect/psql/dm"
 	"github.com/stephenafamo/bob/dialect/psql/im"
 	"github.com/stephenafamo/bob/dialect/psql/um"
+	"github.com/stephenafamo/bob/expr"
 	"github.com/stephenafamo/bob/internal"
 	"github.com/stephenafamo/bob/internal/mappings"
 	"github.com/stephenafamo/bob/orm"
@@ -21,16 +22,16 @@ type (
 	ormDeleteQuery[T any, Tslice ~[]T] = orm.Query[*dialect.DeleteQuery, T, Tslice, bob.SliceTransformer[T, Tslice]]
 )
 
-func NewTable[T any, Tset setter[T]](schema, tableName string) *Table[T, []T, Tset] {
-	return NewTablex[T, []T, Tset](schema, tableName)
+func NewTable[T any, Tset setter[T], C bob.Expression](schema, tableName string, columns C) *Table[T, []T, Tset, C] {
+	return NewTablex[T, []T, Tset](schema, tableName, columns)
 }
 
-func NewTablex[T any, Tslice ~[]T, Tset setter[T]](schema, table string) *Table[T, Tslice, Tset] {
+func NewTablex[T any, Tslice ~[]T, Tset setter[T], C bob.Expression](schema, table string, columns C) *Table[T, Tslice, Tset, C] {
 	setMapping := mappings.GetMappings(reflect.TypeOf((*new(Tset))))
-	view, mappings := newView[T, Tslice](schema, table)
-	t := &Table[T, Tslice, Tset]{
+	view, mappings := newView[T, Tslice](schema, table, columns)
+	t := &Table[T, Tslice, Tset, C]{
 		View:             view,
-		pkCols:           orm.NewColumns(mappings.PKs...).WithParent(view.alias),
+		pkCols:           expr.NewColumnsExpr(mappings.PKs...).WithParent(view.alias),
 		setterMapping:    setMapping,
 		nonGeneratedCols: internal.FilterNonZero(mappings.NonGenerated),
 	}
@@ -40,9 +41,9 @@ func NewTablex[T any, Tslice ~[]T, Tset setter[T]](schema, table string) *Table[
 
 // The table contains extract information from the struct and contains
 // caches ???
-type Table[T any, Tslice ~[]T, Tset setter[T]] struct {
-	*View[T, Tslice]
-	pkCols           orm.Columns
+type Table[T any, Tslice ~[]T, Tset setter[T], C bob.Expression] struct {
+	*View[T, Tslice, C]
+	pkCols           expr.ColumnsExpr
 	setterMapping    mappings.Mapping
 	nonGeneratedCols []string
 
@@ -61,12 +62,12 @@ type Table[T any, Tslice ~[]T, Tset setter[T]] struct {
 }
 
 // Returns the primary key columns for this table.
-func (t *Table[T, Tslice, Tset]) PrimaryKey() orm.Columns {
+func (t *Table[T, Tslice, Tset, C]) PrimaryKey() expr.ColumnsExpr {
 	return t.pkCols
 }
 
 // Starts an insert query for this table
-func (t *Table[T, Tslice, Tset]) Insert(queryMods ...bob.Mod[*dialect.InsertQuery]) *ormInsertQuery[T, Tslice] {
+func (t *Table[T, Tslice, Tset, C]) Insert(queryMods ...bob.Mod[*dialect.InsertQuery]) *ormInsertQuery[T, Tslice] {
 	q := &ormInsertQuery[T, Tslice]{
 		ExecQuery: orm.ExecQuery[*dialect.InsertQuery]{
 			BaseQuery: Insert(im.Into(t.NameAs(), t.nonGeneratedCols...)),
@@ -78,7 +79,7 @@ func (t *Table[T, Tslice, Tset]) Insert(queryMods ...bob.Mod[*dialect.InsertQuer
 	q.Expression.AppendContextualModFunc(
 		func(ctx context.Context, q *dialect.InsertQuery) (context.Context, error) {
 			if !q.HasReturning() {
-				q.AppendReturning(t.Columns())
+				q.AppendReturning(t.Columns)
 			}
 			return ctx, nil
 		},
@@ -90,7 +91,7 @@ func (t *Table[T, Tslice, Tset]) Insert(queryMods ...bob.Mod[*dialect.InsertQuer
 }
 
 // Starts an Update query for this table
-func (t *Table[T, Tslice, Tset]) Update(queryMods ...bob.Mod[*dialect.UpdateQuery]) *ormUpdateQuery[T, Tslice] {
+func (t *Table[T, Tslice, Tset, C]) Update(queryMods ...bob.Mod[*dialect.UpdateQuery]) *ormUpdateQuery[T, Tslice] {
 	q := &ormUpdateQuery[T, Tslice]{
 		ExecQuery: orm.ExecQuery[*dialect.UpdateQuery]{
 			BaseQuery: Update(um.Table(t.NameAs())),
@@ -102,7 +103,7 @@ func (t *Table[T, Tslice, Tset]) Update(queryMods ...bob.Mod[*dialect.UpdateQuer
 	q.Expression.AppendContextualModFunc(
 		func(ctx context.Context, q *dialect.UpdateQuery) (context.Context, error) {
 			if !q.HasReturning() {
-				q.AppendReturning(t.Columns())
+				q.AppendReturning(t.Columns)
 			}
 			return ctx, nil
 		},
@@ -114,7 +115,7 @@ func (t *Table[T, Tslice, Tset]) Update(queryMods ...bob.Mod[*dialect.UpdateQuer
 }
 
 // Starts a Delete query for this table
-func (t *Table[T, Tslice, Tset]) Delete(queryMods ...bob.Mod[*dialect.DeleteQuery]) *ormDeleteQuery[T, Tslice] {
+func (t *Table[T, Tslice, Tset, C]) Delete(queryMods ...bob.Mod[*dialect.DeleteQuery]) *ormDeleteQuery[T, Tslice] {
 	q := &ormDeleteQuery[T, Tslice]{
 		ExecQuery: orm.ExecQuery[*dialect.DeleteQuery]{
 			BaseQuery: Delete(dm.From(t.NameAs())),
@@ -126,7 +127,7 @@ func (t *Table[T, Tslice, Tset]) Delete(queryMods ...bob.Mod[*dialect.DeleteQuer
 	q.Expression.AppendContextualModFunc(
 		func(ctx context.Context, q *dialect.DeleteQuery) (context.Context, error) {
 			if !q.HasReturning() {
-				q.AppendReturning(t.Columns())
+				q.AppendReturning(t.Columns)
 			}
 			return ctx, nil
 		},
@@ -136,3 +137,4 @@ func (t *Table[T, Tslice, Tset]) Delete(queryMods ...bob.Mod[*dialect.DeleteQuer
 
 	return q
 }
+