+++ dialect/sqlite/table.go
@@ -9,6 +9,7 @@ import (
 	"github.com/stephenafamo/bob/dialect/sqlite/dm"
 	"github.com/stephenafamo/bob/dialect/sqlite/im"
 	"github.com/stephenafamo/bob/dialect/sqlite/um"
+	"github.com/stephenafamo/bob/expr"
 	"github.com/stephenafamo/bob/internal/mappings"
 	"github.com/stephenafamo/bob/orm"
 )
@@ -20,16 +21,16 @@ type (
 	ormDeleteQuery[T any, Tslice ~[]T] = orm.Query[*dialect.DeleteQuery, T, Tslice, bob.SliceTransformer[T, Tslice]]
 )
 
-func NewTable[T any, Tset setter[T]](schema, tableName string) *Table[T, []T, Tset] {
-	return NewTablex[T, []T, Tset](schema, tableName)
+func NewTable[T any, Tset setter[T], C bob.Expression](schema, tableName string, columns C) *Table[T, []T, Tset, C] {
+	return NewTablex[T, []T, Tset](schema, tableName, columns)
 }
 
-func NewTablex[T any, Tslice ~[]T, Tset setter[T]](schema, tableName string) *Table[T, Tslice, Tset] {
+func NewTablex[T any, Tslice ~[]T, Tset setter[T], C bob.Expression](schema, tableName string, columns C) *Table[T, Tslice, Tset, C] {
 	setMapping := mappings.GetMappings(reflect.TypeOf(*new(Tset)))
-	view, mappings := newView[T, Tslice](schema, tableName)
-	t := &Table[T, Tslice, Tset]{
+	view, mappings := newView[T, Tslice](schema, tableName, columns)
+	t := &Table[T, Tslice, Tset, C]{
 		View:          view,
-		pkCols:        orm.NewColumns(mappings.PKs...).WithParent(view.alias),
+		pkCols:        expr.NewColumnsExpr(mappings.PKs...).WithParent(view.alias),
 		setterMapping: setMapping,
 	}
 
@@ -38,9 +39,9 @@ func NewTablex[T any, Tslice ~[]T, Tset setter[T]](schema, tableName string) *Ta
 
 // The table contains extract information from the struct and contains
 // caches ???
-type Table[T any, Tslice ~[]T, Tset setter[T]] struct {
-	*View[T, Tslice]
-	pkCols        orm.Columns
+type Table[T any, Tslice ~[]T, Tset setter[T], C bob.Expression] struct {
+	*View[T, Tslice, C]
+	pkCols        expr.ColumnsExpr
 	setterMapping mappings.Mapping
 
 	BeforeInsertHooks bob.Hooks[Tset, bob.SkipModelHooksKey]
@@ -58,12 +59,12 @@ type Table[T any, Tslice ~[]T, Tset setter[T]] struct {
 }
 
 // Returns the primary key columns for this table.
-func (t *Table[T, Tslice, Tset]) PrimaryKey() orm.Columns {
+func (t *Table[T, Tslice, Tset, C]) PrimaryKey() expr.ColumnsExpr {
 	return t.pkCols
 }
 
 // Starts an insert query for this table
-func (t *Table[T, Tslice, Tset]) Insert(queryMods ...bob.Mod[*dialect.InsertQuery]) *ormInsertQuery[T, Tslice] {
+func (t *Table[T, Tslice, Tset, C]) Insert(queryMods ...bob.Mod[*dialect.InsertQuery]) *ormInsertQuery[T, Tslice] {
 	q := &ormInsertQuery[T, Tslice]{
 		ExecQuery: orm.ExecQuery[*dialect.InsertQuery]{
 			BaseQuery: Insert(im.Into(t.NameAs())),
@@ -87,7 +88,7 @@ func (t *Table[T, Tslice, Tset]) Insert(queryMods ...bob.Mod[*dialect.InsertQuer
 }
 
 // Starts an Update query for this table
-func (t *Table[T, Tslice, Tset]) Update(queryMods ...bob.Mod[*dialect.UpdateQuery]) *ormUpdateQuery[T, Tslice] {
+func (t *Table[T, Tslice, Tset, C]) Update(queryMods ...bob.Mod[*dialect.UpdateQuery]) *ormUpdateQuery[T, Tslice] {
 	q := &ormUpdateQuery[T, Tslice]{
 		ExecQuery: orm.ExecQuery[*dialect.UpdateQuery]{
 			BaseQuery: Update(um.Table(t.NameAs())),
@@ -111,7 +112,7 @@ func (t *Table[T, Tslice, Tset]) Update(queryMods ...bob.Mod[*dialect.UpdateQuer
 }
 
 // Starts a Delete query for this table
-func (t *Table[T, Tslice, Tset]) Delete(queryMods ...bob.Mod[*dialect.DeleteQuery]) *ormDeleteQuery[T, Tslice] {
+func (t *Table[T, Tslice, Tset, C]) Delete(queryMods ...bob.Mod[*dialect.DeleteQuery]) *ormDeleteQuery[T, Tslice] {
 	q := &ormDeleteQuery[T, Tslice]{
 		ExecQuery: orm.ExecQuery[*dialect.DeleteQuery]{
 			BaseQuery: Delete(dm.From(t.NameAs())),
@@ -133,3 +134,4 @@ func (t *Table[T, Tslice, Tset]) Delete(queryMods ...bob.Mod[*dialect.DeleteQuer
 
 	return q
 }
+