+++ expr/arg.go
@@ -5,6 +5,7 @@ import (
 	"io"
 
 	"github.com/stephenafamo/bob"
+	"github.com/stephenafamo/bob/internal"
 )
 
 func Arg(vals ...any) bob.Expression {
@@ -45,18 +46,10 @@ func (a args) WriteSQL(ctx context.Context, w io.Writer, d bob.Dialect, start in
 	return a.vals, nil
 }
 
-func toAnySlice[T any](vals ...T) []any {
-	args := make([]any, len(vals))
-	for k, v := range vals {
-		args[k] = v
-	}
-	return args
-}
-
 func ToArgs[T any](vals ...T) bob.Expression {
-	return Arg(toAnySlice(vals...)...)
+	return Arg(internal.ToAnySlice(vals)...)
 }
 
 func ToArgGroup[T any](vals ...T) bob.Expression {
-	return ArgGroup(toAnySlice(vals...)...)
+	return ArgGroup(internal.ToAnySlice(vals)...)
 }