+++ gen/drivers/constraints.go
@@ -59,6 +59,26 @@ type Constraints[Extra any] struct {
 	Checks  []Check[Extra]      `yaml:"check" json:"check"`
 }
 
+func (c Constraints[E]) All() []Constraint[E] {
+	all := make([]Constraint[E], 0, 1+len(c.Foreign)+len(c.Uniques)+len(c.Checks))
+
+	if c.Primary != nil {
+		all = append(all, *c.Primary)
+	}
+
+	for _, fk := range c.Foreign {
+		all = append(all, Constraint[E](fk.Constraint))
+	}
+
+	all = append(all, c.Uniques...)
+
+	for _, check := range c.Checks {
+		all = append(all, Constraint[E](check.Constraint))
+	}
+
+	return all
+}
+
 // Constraint represents a constraint in a database
 type Constraint[Extra any] struct {
 	Name    string   `yaml:"name" json:"name"`