+++ gen/plugins/dberrors.go
@@ -27,7 +27,7 @@ func (dbErrorsPlugin[C]) Name() string {
 
 // PlugState implements gen.StatePlugin.
 func (d dbErrorsPlugin[C]) PlugState(state *gen.State[C]) error {
-	if err := dependsOn(d.config.Disabled, state, "models", "factory"); err != nil {
+	if err := dependsOn(d.config.Disabled, state, "models"); err != nil {
 		return err
 	}