+++ gen/plugins/presets.go
@@ -4,6 +4,7 @@ import "github.com/stephenafamo/bob/internal"
 
 //nolint:gochecknoglobals
 var PresetAll = Config{
+	DBInfo:   OutputConfig{Destination: "dbinfo", Pkgname: "dbinfo"},
 	Enums:    OutputConfig{Destination: "enums", Pkgname: "enums"},
 	Models:   OutputConfig{Destination: "models", Pkgname: "models"},
 	Factory:  OutputConfig{Destination: "factory", Pkgname: "factory"},
@@ -11,11 +12,11 @@ var PresetAll = Config{
 	Where:    OnOffConfig{},
 	Loaders:  OnOffConfig{},
 	Joins:    OnOffConfig{},
-	Names:    OnOffConfig{},
 }
 
 //nolint:gochecknoglobals
 var PresetDefault = Config{
+	DBInfo:   OutputConfig{Destination: "dbinfo", Pkgname: "dbinfo"},
 	Enums:    OutputConfig{Destination: "enums", Pkgname: "enums"},
 	Models:   OutputConfig{Destination: "models", Pkgname: "models"},
 	Factory:  OutputConfig{Destination: "factory", Pkgname: "factory"},
@@ -23,11 +24,11 @@ var PresetDefault = Config{
 	Where:    OnOffConfig{},
 	Loaders:  OnOffConfig{},
 	Joins:    OnOffConfig{},
-	Names:    OnOffConfig{Disabled: internal.Pointer(true)},
 }
 
 //nolint:gochecknoglobals
 var PresetNone = Config{
+	DBInfo:   OutputConfig{Disabled: internal.Pointer(true)},
 	Enums:    OutputConfig{Disabled: internal.Pointer(true)},
 	Models:   OutputConfig{Disabled: internal.Pointer(true)},
 	Factory:  OutputConfig{Disabled: internal.Pointer(true)},
@@ -35,5 +36,4 @@ var PresetNone = Config{
 	Where:    OnOffConfig{Disabled: internal.Pointer(true)},
 	Loaders:  OnOffConfig{Disabled: internal.Pointer(true)},
 	Joins:    OnOffConfig{Disabled: internal.Pointer(true)},
-	Names:    OnOffConfig{Disabled: internal.Pointer(true)},
 }